+++ src/snowflake/snowpark/modin/plugin/extensions/index.py
@@ -855,8 +855,7 @@ class Index(metaclass=TelemetryMeta):
         """
         return self.to_series().any(**kwargs)
 
-    @index_not_implemented()
-    def argmin(self) -> None:
+    def argmin(self, axis=None, skipna: bool = True, *args, **kwargs) -> int:
         """
         Return int position of the smallest value in the Series.
 
@@ -883,11 +882,14 @@ class Index(metaclass=TelemetryMeta):
         numpy.ndarray.argmin : Equivalent method for numpy arrays.
         Series.idxmax : Return index label of the maximum values.
         Series.idxmin : Return index label of the minimum values.
-        """
-        # TODO: SNOW-1458142 implement argmin
 
-    @index_not_implemented()
-    def argmax(self) -> None:
+        Note
+        ----
+        `*args` and `**kwargs` are present for compatibility with numpy and not used with Snowpark pandas.
+        """
+        return self.to_series().argmin(skipna=skipna, *args, **kwargs)
+
+    def argmax(self, axis=None, skipna: bool = True, *args, **kwargs) -> int:
         """
         Return int position of the largest value in the Series.
 
@@ -914,8 +916,12 @@ class Index(metaclass=TelemetryMeta):
         numpy.ndarray.argmax : Equivalent method for numpy arrays.
         Series.idxmax : Return index label of the maximum values.
         Series.idxmin : Return index label of the minimum values.
+
+        Note
+        ----
+        `*args` and `**kwargs` are present for compatibility with numpy and not used with Snowpark pandas.
         """
-        # TODO: SNOW-1458142 implement argmax
+        return self.to_series().argmax(skipna=skipna, *args, **kwargs)
 
     def copy(
         self,